Add search for cars, move all cars EP logic to CarsController, refactor carsRoute
